Feature/#5 메인 화면 UI 및 검색·상세 지도 기능 구현

+++ honglidong/app/src/main/java/com/hongildong/map/ui/search/SearchedItem.kt
@@ -1,4 +1,4 @@
-package com.hongildong.map.ui.home.search
+package com.hongildong.map.ui.search
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
@@ -9,7 +9,6 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.material3.HorizontalDivider
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -25,10 +24,11 @@ import com.hongildong.map.ui.theme.Gray200
 import com.hongildong.map.ui.theme.HongildongTheme
 
 @Composable
-fun RecentlySearchedItem(
+fun SearchedItem(
     itemName: String,
     onClickItem: (String) -> Unit,
-    onDeleteItem: (String) -> Unit
+    onDeleteItem: (String) -> Unit,
+    isRecentlySearched: Boolean = false
 ) {
     Column {
         Row (
@@ -54,16 +54,18 @@ fun RecentlySearchedItem(
                     style = AppTypography.Regular_15
                 )
             }
-            // 삭제 버튼
-            Image(
-                painter = painterResource(R.drawable.ic_close),
-                contentDescription = stringResource(R.string.delete_item, itemName),
-                contentScale = ContentScale.Fit,
-                modifier = Modifier
-                    .padding(4.dp)
-                    .size(12.dp)
-                    .clickable { onDeleteItem(itemName) }
-            )
+            if (isRecentlySearched) {
+                // 삭제 버튼
+                Image(
+                    painter = painterResource(R.drawable.ic_close),
+                    contentDescription = stringResource(R.string.delete_item, itemName),
+                    contentScale = ContentScale.Fit,
+                    modifier = Modifier
+                        .padding(4.dp)
+                        .size(12.dp)
+                        .clickable { onDeleteItem(itemName) }
+                )
+            }
         }
 
         HorizontalDivider(
@@ -78,20 +80,16 @@ fun RecentlySearchedItem(
 fun RecentlySearchedItemPreview() {
     HongildongTheme {
         Column {
-            RecentlySearchedItem(
+            SearchedItem(
                 itemName = "아이템1",
                 onClickItem = {},
                 onDeleteItem = {}
             )
-            RecentlySearchedItem(
+            SearchedItem(
                 itemName = "아이템2",
                 onClickItem = {},
-                onDeleteItem = {}
-            )
-            RecentlySearchedItem(
-                itemName = "아이템3",
-                onClickItem = {},
-                onDeleteItem = {}
+                onDeleteItem = {},
+                isRecentlySearched = true
             )
         }
     }
